Change when stock notif emails are triggered
Moves the check for low/no stock to the point when a product's stock
quantity is updated. This decouples the check from orders, since stock
can also change in other contexts (such as updates via REST API).

Unfortunately this is also causing a bug during classic checkout when
purchasing the last available unit of a product, causing the checkout
screen to show an error, even though the order is still created in the
background. More investigation is still needed here.